+++ lottiefiles/host.php
@@ -6,8 +6,8 @@
  * Description: LottieFiles for WordPress is the easiest way to add Lottie animations to your WordPress website using the Gutenberg editor.
  * Author: LottieFiles
  * Author URI: https://lottiefiles.com/
- * Version: 3.0.0
- * License: GPLv3
+ * Version: 3.1.0
+ * License: GPLv2 or later
  *
  * @package LottieFiles
  */