+++ lottiefiles/src/gutenberg-block/index.tsx
@@ -37,7 +37,7 @@ import { Save } from './frontend/save';
  */
 
 const init = async (): Promise<void> => {
-  const TRACKER_API_KEY = process.env.NODE_ENV === 'development' ? process.env.TRACKER_API_KEY_DEV : process.env.TRACKER_API_KEY;
+  const TRACKER_API_KEY = process.env.TRACKER_API_KEY;
   const TIME = getEpoch();
   const PLATFORM = appDetails.name;
   const APP_VERSION = appDetails.version;